PREFIX SUM 관해서 불판 가는 문제(o) - 해결 -> 해결방법 1,3,6 제거할때 1제거하고 확인 3제거하고 확인 6 제거하고 확인 -> 이런식으로 가서 1제거할때 기준으로는 AppleGrid를 갱신할수 없는 상태인게 맞으니, 제거가 안된것이었음. 1,3,6 한꺼번에 제거하고 난 뒤에 전체적으로 이 apple grid 가 갱신 가능한 상태인지를 봐야 함
불판갈아~ 표지때 컨트롤러 움직일 수 있음(easy) - 해결(o)
restrict 때 startpos 초기화 하기 - 해결 (o)

+++ Assets/WorkThroughout/Scripts/InGame/EndManager/GameEnding.cs
@@ -238,6 +238,7 @@ public class GameEnding : NetworkBehaviour
         // 서버→클라 연장 알림
         NotifyClientsToExtendGameTimeClientRpc();
 
+
        RestrictAllClientsDragClientRpc(seconds);
 
         yield return new WaitForSeconds(extendNoticeDuration);
@@ -246,7 +247,7 @@ public class GameEnding : NetworkBehaviour
 
     //제한 보내는것
     [ClientRpc]
-    private void RestrictAllClientsDragClientRpc(int seconds)
+    private void RestrictAllClientsDragClientRpc(float seconds)
     {
     
         var pc = NetworkManager.Singleton.LocalClient.PlayerObject.GetComponent<PlayerController>();
